添加上传图片、添加评论功能。

+++ controllers/image.js
@@ -1,44 +1,42 @@
 let fs = require('fs'),
   path = require('path'),
-  sidebar = require('../helpers/sidebar.js')
+  sidebar = require('../helpers/sidebar.js'),
+  Models = require('../models'),
+  md5 = require('MD5')
 
 module.exports = {
   index: (req, res) => {
     // res.send(`the image: index controller ${req.params.image_id}`)
     // res.render('image')
     let viewModel = {
-      image: {
-        uniqueId: 1,
-        title: 'Sample Image 1',
-        description: 'This is a sample',
-        filename: 'sample1.jpg',
-        views: 0,
-        likes: 0,
-        timestamp:  Date.now()
-      },
-      comments: [
-        {
-          image_id: 1,
-          email: 'test@testing.com',
-          name: 'Test Tester',
-          gravatar: 'http://lorempixel.com/75/75/animals/1',
-          comment: 'This is a test comment...',
-          timestamp: Date.now()
-        },
-        {
-          image_id: 2,
-          email: 'test@testing.com',
-          name: 'Test Tester',
-          gravatar: 'http://lorempixel.com/75/75/animals/2',
-          comment: 'This is a test comment...',
-          timestamp: Date.now()
-        },
-      ]
     }
     // let viewModel = {}
     // res.render('image', viewModel)
-    sidebar(viewModel, function (viewModel) {
-      res.render('image', viewModel)
+    // sidebar(viewModel, function (viewModel) {
+    //   res.render('image', viewModel)
+    // })
+    Models.Image.findOne({
+      filename: { $regex: req.params.image_id}
+    }, function (err, image) {
+      if (err) {throw err}
+      if (image) {
+        image.views++
+        viewModel.image = image
+        image.save()
+        Models.Comment.find({image_id: image.image_id}, {}, {
+        // Models.Comment.find({image_id: image.image_id}, {}, {
+          sort: {
+            timestamp: -1
+          }
+        }, function (err, comments) {
+          viewModel.comments = comments
+          sidebar(viewModel, (viewModel) => {
+            res.render('image', viewModel)
+          })
+        })
+      } else {
+        res.redirect('/')
+      }
     })
   },
   create: (req, res) => {
@@ -50,28 +48,67 @@ module.exports = {
       for (let i = 0; i < 6; i++) {
         imgUrl += possible.charAt(Math.floor(Math.random() * possible.length))
       }
-      let tempPath = req.files[0].path,
-        ext = path.extname(req.files[0].originalname).toLowerCase(),
-        targetPath = path.resolve('./public/upload/' + imgUrl + ext)
-      if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
-        fs.rename(tempPath, targetPath, function (err) {
-          if (err) throw err
-          res.redirect('/images/' + imgUrl)
-        })
-      } else {
-        fs.unlink(tempPath, () => {
-          if (err) throw err;
-          res.json(500, {error: 'Only image files are allowed.'})
-        })
-      }
+      Models.Image.find({filename: imgUrl}, function (err, images) {
+        if (images.length > 0) {
+          saveImage()
+        } else {
+          let tempPath = req.files[0].path,
+            ext = path.extname(req.files[0].originalname).toLowerCase(),
+            targetPath = path.resolve('./public/upload/' + imgUrl + ext)
+          if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
+            fs.rename(tempPath, targetPath, function (err) {
+              if (err) {throw err}
+              var newImg = new Models.Image({
+                title: req.body.title,
+                filename: imgUrl + ext,
+                image_id: imgUrl,
+                description: req.body.description
+              })
+              newImg.save(function(err, image) {
+                res.redirect('/images/' + image.uniqueId)
+              })
+            })
+          } else {
+            fs.unlink(tempPath, () => {
+              if (err) { throw err }
+              res.json(500, {error: 'Only image files are allowed.'})
+            })
+          }
+        }
+      })
     }
     saveImage()
   },
   like: function (req, res) {
     // res.send('the image: like POST controller')
-    res.json({likes: 1})
+    // res.json({likes: 1})
+    Models.Image.findOne({filename: {$regex: req.params.image_id}}, function (err, image) {
+      if (!err && image) {
+        image.likes++
+        image.save(function (err) {
+          if (err) {
+            res.json(err)
+          } else {
+            res.json({likes: image.likes})
+          }
+        })
+      }
+    })
   },
   comment: function (req, res) {
-    res.send('the image: comment POST controller')
+    // res.send('the image: comment POST controller')
+    Models.Image.findOne({filename: {$regex: req.params.image_id}}, function(err, image) {
+      if (!err && image) {
+        let newComment = new Models.Comment(req.body)
+        newComment.gravatar = md5(newComment.email)
+        newComment.image_id = req.params.image_id
+        newComment.save(function(err, comment) {
+          if (err) {throw err}
+          res.redirect(`/images/${image.uniqueId}#${comment._id}`)
+        })
+      } else {
+        res.redirect('/')
+      }
+    })
   }
 }
